+++ src/Components/IdentityManagement.js
@@ -2,11 +2,16 @@ import React from "react";
 import { connect } from "react-redux";
 
 import { nameList, wallet } from "../selectors";
-import { UsernameList, GhostButton, GoIcon, ActionButton } from "../Styles";
+import {
+  SelectableList,
+  SelectableItem,
+  GhostButton,
+  GoIcon,
+  ActionButton
+} from "../Styles";
 import IdentityCredit from "./IdentityCredit";
 
 function IdentityManagement({ names, index, select, createUsername, hide }) {
-  console.log(names, index);
   return (
     <>
       <h2>Identity Management</h2>
@@ -20,20 +25,19 @@ function IdentityManagement({ names, index, select, createUsername, hide }) {
         button.
       </p>
       <h3>Select a username</h3>
-      <UsernameList>
+      <SelectableList>
         {names.map(name => {
-          console.log(name);
           return (
-            <li>
+            <SelectableItem isSelected={name.isSelected}>
               <GhostButton onClick={e => select(name.name, index) || hide()}>
                 {name.name}
                 <IdentityCredit>{name.balance}</IdentityCredit>
                 <GoIcon />
               </GhostButton>
-            </li>
+            </SelectableItem>
           );
         })}
-      </UsernameList>
+      </SelectableList>
       <ActionButton onClick={createUsername} ownRow={true}>
         Create Username
       </ActionButton>
@@ -55,7 +59,6 @@ const dispatchToProps = dispatch => {
     },
     createUsername() {
       dispatch({ type: "SHOW_CREATE_USERNAME" });
-      dispatch({ type: "HIDE_IDENTITY_MANAGEMENT" });
     },
     hide() {
       dispatch({ type: "HIDE_IDENTITY_MANAGEMENT" });